Minor changes and fixes are made
The message that appears when the "Add To Do" button is clicked while the entry areas in "To Do Frame" are empty is updated.

The problem that the main window does not change its dimensions when new activity is added. The windows dimension specifiers in the geometry method are removed.

The "spent_time/total_time" is changed to minutes as it should be from seconds when an activity is ended by clicking "Done" button.

The name of TodayPomodoro is changed to TodayActivityFrame and other variables and class name are changed.